Add AddressBook wrapper
This adds a simple wrapper struct around ABAddressBook, which is
straight up horrendous to use natively in Swift.

Usage: call `AddressBook.getAddressBook` to try and get the user's
address book. If we need to ask the user for permission, we will.
Assuming the user grants access, we will pass a Result with the
instantiated AddressBook in the Success case. If we fail for any reason
(disallowed access, general errors), we will get an empty Failure case.

Once you have an `AddressBook` instance, you can access the
`localPeople` property to get a list of `LocalPerson` instances. These
objects are simple and have a `name` and a list of associated email
addresses.