fixes the omnibar padding
The omnibar padding was wrong, now it is lovely. The
UITextView's scrolling inset is set to a *negative*
value on the right side, and so the entire view has
its clipsToBounds property as 'false'.  Usually this
would cause the text to display when scrolling, so
the parent container view has *its* clipsToBounds
set to 'true' to compensate.  Very goofy, if you
ask me!
[Finishes: #91898692]