Tweak ElloLabel sizing to remove temporal coupling
Previously, we had to make sure that ElloLabel instances had a non-zero
frame before getting their height. This caused issues where we'd have to
make sure we added a width and called `sizeToFit` before accessing the
height. Instead of that, we can rework the existing implementation to
take a desired width and use that to perform the proper calculations.
That lets us pass the tableView frame's width and let UIKit handle the
actual sizing of the label.