Resize the alert on small screens
If the alert is too large for the screen size, we will use a default
height of screen size - padding, and use that for the desired size. Once
we've displayed the alert, if we are using this size, we will enable
scrolling so that users can access all of the buttons.