Force profile header cell to display with the correct height
This one sucks. `ProfileViewController` is loaded via two mechanisms. One way is to tap on a user's avatar in a stream. This case sizes the content correctly because we have to load the user from the api so there is a delay before the view is rendered. The second way is directly from the main nav. In this case we have the profile of the current user and can display it immediately. The issue is that the `ProfileViewController` and it's child `StreamViewController` are not yet at their proper size when we attempt to calculate the height of the profile content. Both views are 600px wide. The solution, a hack, is to wait for `viewWillAppear` on `ProfileViewController` and then pass the size into `appendUnsizedCellItems(_:withWidth:)` on dataSource.

[Fixes #92366706]