Prevent small images from scaling to the full width of the screen
The solution (for now) is to scale an image up to it's non-retina width. For example, on the iPhone 6 (375pts wide) we will display a 375px width image at full screen width. Images smaller than the width of the screen (pixel width image, pt width screen) will not scale past their pixel width in screen pts.

[Fixes #92263260]