LOOK at THIS - ALL of SEARCH
It's such a tiny commit!  And yeah, all
of user search in this one commit. Thanks
to the awesome design of the StreamViewController
(@steam) and the sick 'UserList' StreamKind,
plus infinite scroll being baked into everything
(@rynbyjn and api team).  All I did was attach
rod A to slot B, and wrote this commit message.

But hey, this commit message WAS a bit of work!
It didn't just spring forth from nothing.  I had
to type the darn thing.  And I write all my
commit messages in the shell, so if there was a
typo, I only had one chance to fix it (the current
line that is).  I hope my '<50 chars' approximation
is close, I'm just guessing at it.

Peace out!