Add account deletion to the settings
This adds a delete button to a dynamic setting cell. The presenter is
used to decide if the delete button is showing or hidden. There is a
hard coded category and setting for account deletion to make it work
with the other dynamic settings. Before the account is deleted, we ask
the user one more time for confirmation.