Add PushNotificationController
This class handles the specifics around registering for push
notifications.

We call `registerForRemoteNotifications()` _a lot_. We call this method
every time the app launches, as well as any time we log in.

If the user isn't currently authenticated, we don't continue.

If the user is authenticated, the first thing we do is try to upload the
stored token. This is because the API for registering for push
notifications might not return a new token, and so we'll never get a
callback.

After we've tried to upload the existing token, we ask for permission.
If the user hasn't granted permission already, this will present the
standard system dialog. Once they accept, we will pass the token from
the app delegate to the `updateToken` method, save it to the keychain,
then update the current user's token with the server.

If we get either of the logged out notifications, we remove the token
from the server (for the current user), but we leave the token in the
keychain. This is crucial, since Apple might never give us another
token.