Prevent credential settings from auto scrolling
When the keyboard opens from a UITextView becoming first responder, the
table view is automatically scrolled to make the text view visibile if
the keyboard was going to open over it. This prevents that from
happening on this table view and allows it to happen on the parent table
view.