Disable Swift interaction tracing to prevent crash
The app is crashing consistently when you go to change your password in
the settings and navigate away from the app and to 1Password. When you
come back into the app it crashes with an exception tracing back into
NRMAHarvester. Turning off this feature of New Relic fixes this crash.