Enable save button when password is valid
This fixes 2 bugs. One bug was that the function being called when
either password field had changed, had the same name. The second was
that the observer of the current password field was only being set in
the callback function of the username field change observer. This fixes
both of those.