Properly set the selected tab bar item
We can't set the tab bar's selected item to the tab bar item of the view
controller we're trying to show because those are actually different
instances. Instead, we need to get the tab bar item at the same index as
the incoming view controller.

Also moving the code for this to `showViewController`, since it makes
more semantic sense to have that be a side effect of showing a view
controller.