Use push payload to determine selected tab
URLs and push payloads now exhibit the same behavior.

When the app is launched via url, we strip the protocol (`ello://`) and
pass the rest of the url's absolute string value to the
`navigateToDeepLink` method (previously named `handleURL`) on the
`AppViewController`. This method shows the proper tab, and can be
expanded to handle showing individual items.

When the app is brought to the foreground via push notification, we pull
the `"application_target"` key from the payload, which follows the same
format as the url schemes and pass that to the `navigateToDeepLink`
method.