Refresh the notifications screen when navigating
If we receive a push notification, or are launched via URL, we should
reload the Notifications tab to ensure that we have the newest info. In
order to do this, we can post a new local notification from the app view
controller indicating that we're going to show that screen, and then the
notifications view controller can refresh itself accordingly.