hopefully this fixes the friggin logout issue
What @rynbyjn discovered is that when two
requests go out at the same time, and the
auth token is expired, both will request a
refresh token.  One of them (the first)
will succeed, the other gets a 401.

This is a 'very quick fix' solution: if the
token request failed, but the current token
is *different* (and valid), then try again.