UI updates related to the new 'TemporaryCache'
The cache is used in the Profile and Settings screens,
and this code also uses the 'refreshed user' info
to assign an up-to-date avatar and coverImage.  So if
the user uploads these images from the web, they can
be updated in the app using pull-to-refresh.

In Settings, newly uploaded images were not being
assigned to the appropriate image views, which meant
those images would actually *never* be updated (since
the avatar and cover image assets were never being
updated).