Toggle the status bar when showing streams
The drawer makes this a weird case, but we need to show or hide the
status bar when transitioning from the drawer to the streams view
controller. Since the drawer _always_ shows the status bar, we need to
make sure to correct the visibility of the status bar based on whether
or not we're currently showing the navigation bar.