Modify stream container to show the status bar
The main stream view controllers were really misbehaving with this
change.

- When showing a new stream, we need to set all of the streams'
  `scrollsToTop` to `false` in order to prevent them from conflicting.
  Then we can set `scrollsToTop` to `true` for the stream we're showing.
- Modify the layout and logic around showing/hiding the nav bars to
  behave.

Using the default `updateInsets` behavior provided by
`StreamableViewController` wasn't working properly here. Two things were
happening:

1. _Something_ (still not sure what) was pushing the collection views
   down by 20 points because of the status bar. This meant that there was
   too much padding between the top cell and the nav bar when the nav bar
   was visible.
2. When we updated the offset in response to the hiding/showing of the
   navigation bars, the offscreen stream didn't have a nav bar to push off
   of, and so was being mis-set so that when you switched to it, it was
   tucked under the navigation bar. It took a while to see what was
   happening because issue #1 meant the math was really weird.

The fix to this was actually fairly simple:

1. Instead of tying the top of the scollview's frame to the top layout
   guide, we tie it to the bottom of the nav bar.
2. Instead of using the `updateInsets` method provided by
   `StreamableViewController`, we ignore the insets entirely and simply
   trigger a layout pass with `layoutIfNeeded` after updating the
   navigation bar's frame.

This fixes the issues by completely bypassing both of them. We're no
longer at risk of being pushed down by the status bar, because we aren't
relying on it. And because we're modifying the _frame_ of the scrollview
instead of the content insets, the behavior is consistent between both
the on-screen and off-screen stream view controllers, no matter how many
we have.

We _aren't_ animating the layout pass, but I wasn't able to see any
visual issues around this. This also means we aren't updating the bottom
content inset, but I wasn't able to see any visual bugs around this
either.