fixes the weird animations
- in notifications (and other places), because of setStatusBarHidden again
- in StreamContainerViewController, because of layoutIfNeeded

removed animations from setStatusBarHidden

for StreamContainerViewController, the reason it
couldn't use our usual 'updateInsets' method is
because it was being tied to the `topLayoutGuide`,
instead of the superview.top

adding the status bar pushed the top of the
scrollview by 20pt.  the insets, though, were being
based on the navigationBar *height*, and didn't
take the status bar height into account.

using the correct constraint, I could restore the
old updateInsets code (which also removed the
animation bug).