Add SSL Pinning support to the app.
Still need to:
- [ ] Create a valid `.cer` file
- [ ] Add valid `.cer` to `Resources`
- [ ] Decide on how we update the certificate

Due to the nature of the certs we will need to update them as often as we see fit. Google does this every month, but seems like most do it every 1-2 years. This will require us to update the app as it is in current form. We could also write a way for the app to be able to update the local `.cer` from an API endpoint.

[#90556544]